:art:
Improving structure
