+++ js/main.js
@@ -1,6 +1,6 @@
-import intro from '../templates/intro';
+import intro from './templates/intro';
 
-import draw from '../js/draw.js';
+import draw from './draw.js';
 
 (() => {
   draw(intro);
