+++ js/templates/greeting.js
@@ -1,7 +1,7 @@
-import getElementFromTemplate from '../js/getElementFromTemplate';
+import getElementFromTemplate from '../getElementFromTemplate';
 
-import rules from '../templates/rules';
-import draw from '../js/draw';
+import rules from './rules';
+import draw from '../draw';
 
 const greeting = getElementFromTemplate(`<div class="greeting  central--blur">
     <div class="greeting__logo"><img src="img/logo_big.png" width="201" height="89" alt="Pixel Hunter"></div>
@@ -17,9 +17,9 @@ const greeting = getElementFromTemplate(`<div class="greeting  central--blur">
     <div class="greeting__continue"><span><img src="img/arrow_right.svg" width="64" height="64" alt="Next"></span></div>
   </div>`);
 
-const greeting__continue = greeting.querySelector('.greeting__continue');
+const greetingContinue = greeting.querySelector('.greeting__continue');
 
 const handler = (e) => draw(rules);
-greeting__continue.addEventListener('click', handler);
+greetingContinue.addEventListener('click', handler);
 
 export default greeting;
